Slave

+++ src/main/java/pl/wallet/Wallet.java
@@ -1,11 +1,15 @@
 package pl.wallet;
 
-import lombok.*;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import pl.user.User;
 import pl.wallet.transaction.Transaction;
 
 import javax.persistence.*;
 import javax.transaction.Transactional;
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 
@@ -26,7 +30,7 @@ public class Wallet {
   private String name;
 
   @Column(nullable = false)
-  private Long balance;
+  private BigDecimal balance;
 
   @OneToMany
   private List<Transaction> transactions;
